feat(cxl-ui): rename jw player component
