+++ packages/cxl-ui/scss/cxl-jw-player/chapter.scss
