+++ packages/cxl-ui/src/components/cxl-jw-player/README.md
@@ -1,15 +1,15 @@
-# JW Player
+# CXL JW Player
 
 ## Usage
 
 ```
-    <jw-player
+    <cxl-jw-player
       captions
       media-id="fZ0XiGdb"
       minimum-search-length="3"
       player-id="5CFJNXKb"
       plugin-path="https://cxl.com/institute/wp-content/plugins/cxl-jwplayer/"
-    ></jw-player>
+    ></cxl-jw-player>
 ```
 
 ## Features:
