--- packages/cxl-ui/src/styles/jw-player/.gitignore
@@ -1,1 +1,0 @@
-!.gitignore
