+++ packages/storybook/cxl-ui/cxl-jw-player/index.stories.js
@@ -1,8 +1,8 @@
 import { html } from 'lit';
-import '@conversionxl/cxl-ui/src/components/jw-player/index.js';
+import '@conversionxl/cxl-ui/src/components/cxl-jw-player/index.js';
 
 export default {
-  title: 'JW Player/JW Player',
+  title: 'CXL UI/cxl-jw-player',
 };
 
 const Template = ({ captions, mediaId, minimumSearchLength, playerId, playlistId, pluginPath }) =>
@@ -12,23 +12,23 @@ const Template = ({ captions, mediaId, minimumSearchLength, playerId, playlistId
         height: 100vh;
       }
 
-      jw-player {
+      cxl-jw-player {
         display: block;
         height: 100%;
       }
 
-      jw-player * {
+      cxl-jw-player * {
         box-sizing: border-box;
       }
     </style>
-    <jw-player
+    <cxl-jw-player
       ?captions=${captions}
       media-id=${mediaId}
       minimum-search-length=${minimumSearchLength}
       player-id=${playerId}
       playlist-id=${playlistId}
       plugin-path="${pluginPath}"
-    ></jw-player>
+    ></cxl-jw-player>
   `;
 
 export const Default = Template.bind({});
